% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_api_version}
\alias{get_api_version}
\title{Get the version of iPTMnet API being used.}
\usage{
get_api_version()
}
\value{
A string representing the version being used.
}
\description{
Get the version of the api that is being used by the client for making the requests.
}
\examples{
version <- get_api_version()
}
