% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_host_url}
\alias{get_host_url}
\title{Get the url of iPTMnet API server.}
\usage{
get_host_url()
}
\value{
A string representing the URL.
}
\description{
Get the URl that is being used by the client for making the requests.
}
\examples{
url <- get_host_url()
}
