% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_info}
\alias{get_info}
\title{Get information.}
\usage{
get_info(id)
}
\arguments{
\item{id}{A string representing iPTMnet ID}
}
\value{
A list containing the information for the iPTMnet ID.
}
\description{
Get the information for given iPTMnet ID.
}
\examples{
info <- get_info("Q15796")
}
