% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_ppi_for_proteoforms}
\alias{get_ppi_for_proteoforms}
\title{Get Protein-Protein interactions along with corresponding proteoforms.}
\usage{
get_ppi_for_proteoforms(id)
}
\arguments{
\item{id}{A string representing iPTMnet ID.}
}
\value{
A dataframe containing the PTm dependent PPI for the given iPTMnet ID.
}
\description{
Get Protein-Protein interactions along with corresponding proteoforms for the given iPTMnet ID.
}
\examples{
ppi_proteoforms <- get_ppi_for_proteoforms("Q15796")
}
