% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_proteoforms}
\alias{get_proteoforms}
\title{Get proteoforms}
\usage{
get_proteoforms(id)
}
\arguments{
\item{id}{A string representing iPTMnet ID.}
}
\value{
A dataframe containing the proteoforms for the given iPTMnet ID.
}
\description{
Get proteoforms
}
\examples{
proteoforms <- get_proteoforms("Q15796")
}
