% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_ptm_dependent_ppi}
\alias{get_ptm_dependent_ppi}
\title{Get Post translational modification(PTM) dependent Protein-Protein interactions.}
\usage{
get_ptm_dependent_ppi(id)
}
\arguments{
\item{id}{A string representing iPTMnet ID.}
}
\value{
A dataframe containing the PTm dependent PPI for the given iPTMnet ID.
}
\description{
Get Post translational modification(PTM) dependent Protein-Protein interactions for the given iPTMnet ID.
}
\examples{
ptm_dependent_ppi <- get_ptm_dependent_ppi("Q15796")
}
