% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_substrates}
\alias{get_substrates}
\title{Get substrates.}
\usage{
get_substrates(id)
}
\arguments{
\item{id}{A string representing iPTMnet ID.}
}
\value{
A dataframe containing the substrates for the given iPTMnet ID.
}
\description{
Retrieves the substrates for the given iPTMnet ID.
}
\examples{
substrates <- get_substrates("Q15796")
}
