% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_variants}
\alias{get_variants}
\title{Get variants.}
\usage{
get_variants(id)
}
\arguments{
\item{id}{A string representing iPTMnet ID.}
}
\value{
A dataframe containing the variants for the given iPTMnet ID.
}
\description{
Get variants for the given iPTMnet ID.
}
\examples{
variants <- get_variants("Q15796")
}
