% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{PTMTypes}
\alias{PTMTypes}
\title{PTMTypes
Returns a list of strings that represents the PTM types to be used in \link{search_iptmnet} function}
\usage{
PTMTypes()
}
\value{
A list of strings representing PTM types
}
\description{
PTMTypes
Returns a list of strings that represents the PTM types to be used in \link{search_iptmnet} function
}
\examples{
PTMTypes()$Acetylation
PTMTypes()$CGlycosylation
PTMTypes()$Myristoylation
PTMTypes()$Ubiquitination
PTMTypes()$NGlycosylation
PTMTypes()$SGlycosylation
PTMTypes()$Phosphorylation
PTMTypes()$SNitrosylation
PTMTypes()$OGlycosylation
PTMTypes()$Methylation
PTMTypes()$Sumoylation
}
