% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Role}
\alias{Role}
\title{Role
Returns a list of strings that represent the role to be used in \link{search_iptmnet} function}
\usage{
Role()
}
\value{
A list of strings representing the role
}
\description{
Role
Returns a list of strings that represent the role to be used in \link{search_iptmnet} function
}
\examples{
Role()$EnzymeOrSubstrate
Role()$Enzyme
Role()$Substrate
Role()$EnzymeAndSubstrate
}
