% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_api_version}
\alias{set_api_version}
\title{Set the version of the API to use.}
\usage{
set_api_version(version)
}
\arguments{
\item{version}{A string representing the new iTPMnet api version.}
}
\description{
This function can be used to change the version of the api to use.
}
\examples{
set_api_version("v1")
}
