% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_host_url}
\alias{set_host_url}
\title{Set the url of iPTMnet API server.}
\usage{
set_host_url(url)
}
\arguments{
\item{url}{A string representing the new iTPMnet api server URL.}
}
\description{
This function can be used to change the URL to use your own self hosted instance of iPTMnet API server.
}
\examples{
set_host_url("http://www.example.com")
}
