% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{TermType}
\alias{TermType}
\title{TermTypes
Returns a list of strings that represent the term types to be used in \link{search_iptmnet} function.}
\usage{
TermType()
}
\value{
A list of strings that represent the term types.
}
\description{
TermTypes
Returns a list of strings that represent the term types to be used in \link{search_iptmnet} function.
}
\examples{
TermType()$ALL
TermType()$UniprotID
}
